feat: move api destination and connection tests to separate test file
