+++ tests/aws/services/events/test_events.py
@@ -2,7 +2,6 @@
 Test creating and modifying event buses, as well as putting events to custom and the default bus.
 """
 
-import base64
 import datetime
 import json
 import os
@@ -13,18 +12,15 @@ import uuid
 import pytest
 from botocore.exceptions import ClientError
 from localstack_snapshot.snapshots.transformer import SortingTransformer
-from pytest_httpserver import HTTPServer
-from werkzeug import Request, Response
 
 from localstack import config
 from localstack.services.events.v1.provider import _get_events_tmp_dir
 from localstack.testing.aws.eventbus_utils import allow_event_rule_to_sqs_queue
 from localstack.testing.aws.util import is_aws_cloud
 from localstack.testing.pytest import markers
-from localstack.testing.snapshots.transformer_utility import TransformerUtility
 from localstack.utils.files import load_file
-from localstack.utils.strings import long_uid, short_uid, to_str
-from localstack.utils.sync import poll_condition, retry
+from localstack.utils.strings import long_uid, short_uid
+from localstack.utils.sync import retry
 from tests.aws.services.events.helper_functions import (
     assert_valid_event,
     is_old_provider,
@@ -50,32 +46,6 @@ TEST_EVENT_PATTERN_NO_SOURCE = {
     "detail": {"command": ["update-account"]},
 }
 
-API_DESTINATION_AUTHS = [
-    {
-        "type": "BASIC",
-        "key": "BasicAuthParameters",
-        "parameters": {"Username": "user", "Password": "pass"},
-    },
-    {
-        "type": "API_KEY",
-        "key": "ApiKeyAuthParameters",
-        "parameters": {"ApiKeyName": "Api", "ApiKeyValue": "apikey_secret"},
-    },
-    {
-        "type": "OAUTH_CLIENT_CREDENTIALS",
-        "key": "OAuthParameters",
-        "parameters": {
-            "AuthorizationEndpoint": "replace_this",
-            "ClientParameters": {"ClientID": "id", "ClientSecret": "password"},
-            "HttpMethod": "put",
-            "OAuthHttpParameters": {
-                "BodyParameters": [{"Key": "oauthbody", "Value": "value1"}],
-                "HeaderParameters": [{"Key": "oauthheader", "Value": "value2"}],
-                "QueryStringParameters": [{"Key": "oauthquery", "Value": "value3"}],
-            },
-        },
-    },
-]
 
 EVENT_BUS_ROLE = {
     "Statement": {
@@ -282,165 +252,6 @@ class TestEvents:
         )
 
         assert [json.loads(event["Detail"]) for event in sorted_events] == event_details_to_publish
-
-    @markers.aws.only_localstack
-    @pytest.mark.parametrize("auth", API_DESTINATION_AUTHS)
-    def test_api_destinations(self, httpserver: HTTPServer, auth, aws_client, clean_up):
-        token = short_uid()
-        bearer = f"Bearer {token}"
-
-        def _handler(_request: Request):
-            return Response(
-                json.dumps(
-                    {
-                        "access_token": token,
-                        "token_type": "Bearer",
-                        "expires_in": 86400,
-                    }
-                ),
-                mimetype="application/json",
-            )
-
-        httpserver.expect_request("").respond_with_handler(_handler)
-        http_endpoint = httpserver.url_for("/")
-
-        if auth.get("type") == "OAUTH_CLIENT_CREDENTIALS":
-            auth["parameters"]["AuthorizationEndpoint"] = http_endpoint
-
-        connection_name = f"c-{short_uid()}"
-        connection_arn = aws_client.events.create_connection(
-            Name=connection_name,
-            AuthorizationType=auth.get("type"),
-            AuthParameters={
-                auth.get("key"): auth.get("parameters"),
-                "InvocationHttpParameters": {
-                    "BodyParameters": [
-                        {
-                            "Key": "connection_body_param",
-                            "Value": "value",
-                            "IsValueSecret": False,
-                        },
-                    ],
-                    "HeaderParameters": [
-                        {
-                            "Key": "connection-header-param",
-                            "Value": "value",
-                            "IsValueSecret": False,
-                        },
-                        {
-                            "Key": "overwritten-header",
-                            "Value": "original",
-                            "IsValueSecret": False,
-                        },
-                    ],
-                    "QueryStringParameters": [
-                        {
-                            "Key": "connection_query_param",
-                            "Value": "value",
-                            "IsValueSecret": False,
-                        },
-                        {
-                            "Key": "overwritten_query",
-                            "Value": "original",
-                            "IsValueSecret": False,
-                        },
-                    ],
-                },
-            },
-        )["ConnectionArn"]
-
-        # create api destination
-        dest_name = f"d-{short_uid()}"
-        result = aws_client.events.create_api_destination(
-            Name=dest_name,
-            ConnectionArn=connection_arn,
-            InvocationEndpoint=http_endpoint,
-            HttpMethod="POST",
-        )
-
-        # create rule and target
-        rule_name = f"r-{short_uid()}"
-        target_id = f"target-{short_uid()}"
-        pattern = json.dumps({"source": ["source-123"], "detail-type": ["type-123"]})
-        aws_client.events.put_rule(Name=rule_name, EventPattern=pattern)
-        aws_client.events.put_targets(
-            Rule=rule_name,
-            Targets=[
-                {
-                    "Id": target_id,
-                    "Arn": result["ApiDestinationArn"],
-                    "Input": '{"target_value":"value"}',
-                    "HttpParameters": {
-                        "PathParameterValues": ["target_path"],
-                        "HeaderParameters": {
-                            "target-header": "target_header_value",
-                            "overwritten_header": "changed",
-                        },
-                        "QueryStringParameters": {
-                            "target_query": "t_query",
-                            "overwritten_query": "changed",
-                        },
-                    },
-                }
-            ],
-        )
-
-        entries = [
-            {
-                "Source": "source-123",
-                "DetailType": "type-123",
-                "Detail": '{"i": 0}',
-            }
-        ]
-        aws_client.events.put_events(Entries=entries)
-
-        # clean up
-        aws_client.events.delete_connection(Name=connection_name)
-        aws_client.events.delete_api_destination(Name=dest_name)
-        clean_up(rule_name=rule_name, target_ids=target_id)
-
-        to_recv = 2 if auth["type"] == "OAUTH_CLIENT_CREDENTIALS" else 1
-        poll_condition(lambda: len(httpserver.log) >= to_recv, timeout=5)
-
-        event_request, _ = httpserver.log[-1]
-        event = event_request.get_json(force=True)
-        headers = event_request.headers
-        query_args = event_request.args
-
-        # Connection data validation
-        assert event["connection_body_param"] == "value"
-        assert headers["Connection-Header-Param"] == "value"
-        assert query_args["connection_query_param"] == "value"
-
-        # Target parameters validation
-        assert "/target_path" in event_request.path
-        assert event["target_value"] == "value"
-        assert headers["Target-Header"] == "target_header_value"
-        assert query_args["target_query"] == "t_query"
-
-        # connection/target overwrite test
-        assert headers["Overwritten-Header"] == "original"
-        assert query_args["overwritten_query"] == "original"
-
-        # Auth validation
-        match auth["type"]:
-            case "BASIC":
-                user_pass = to_str(base64.b64encode(b"user:pass"))
-                assert headers["Authorization"] == f"Basic {user_pass}"
-            case "API_KEY":
-                assert headers["Api"] == "apikey_secret"
-
-            case "OAUTH_CLIENT_CREDENTIALS":
-                assert headers["Authorization"] == bearer
-
-                oauth_request, _ = httpserver.log[0]
-                oauth_login = oauth_request.get_json(force=True)
-                # Oauth login validation
-                assert oauth_login["client_id"] == "id"
-                assert oauth_login["client_secret"] == "password"
-                assert oauth_login["oauthbody"] == "value1"
-                assert oauth_request.headers["oauthheader"] == "value2"
-                assert oauth_request.args["oauthquery"] == "value3"
 
     @markers.aws.validated
     @pytest.mark.skipif(is_old_provider(), reason="V1 provider does not support this feature")
@@ -1808,383 +1619,4 @@ class TestEventTarget:
             Targets=[
                 {"Id": target_id, "Arn": queue_arn, "InputPath": "$.detail"},
             ],
-        )
-
-
-API_DESTINATION_AUTH_PARAMS = [
-    {
-        "AuthorizationType": "BASIC",
-        "AuthParameters": {
-            "BasicAuthParameters": {"Username": "user", "Password": "pass"},
-        },
-    },
-    {
-        "AuthorizationType": "API_KEY",
-        "AuthParameters": {
-            "ApiKeyAuthParameters": {"ApiKeyName": "ApiKey", "ApiKeyValue": "secret"},
-        },
-    },
-    {
-        "AuthorizationType": "OAUTH_CLIENT_CREDENTIALS",
-        "AuthParameters": {
-            "OAuthParameters": {
-                "AuthorizationEndpoint": "https://example.com/oauth",
-                "ClientParameters": {"ClientID": "client_id", "ClientSecret": "client_secret"},
-                "HttpMethod": "POST",
-            }
-        },
-    },
-]
-
-
-class TestEventBridgeConnections:
-    @pytest.fixture(autouse=True)
-    def connection_snapshots(self, snapshot, connection_name):
-        """Common snapshot transformers for connection tests."""
-        return TransformerUtility.eventbridge_connection(snapshot, connection_name)
-
-    @markers.aws.validated
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    def test_create_connection(self, aws_client, snapshot, create_connection, connection_name):
-        response = create_connection(
-            "API_KEY",
-            {
-                "ApiKeyAuthParameters": {"ApiKeyName": "ApiKey", "ApiKeyValue": "secret"},
-                "InvocationHttpParameters": {},
-            },
-        )
-        snapshot.match("create-connection", response)
-
-        describe_response = aws_client.events.describe_connection(Name=connection_name)
-        snapshot.match("describe-connection", describe_response)
-
-    @markers.aws.validated
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    @pytest.mark.parametrize("auth_params", API_DESTINATION_AUTH_PARAMS)
-    def test_create_connection_with_auth(
-        self, aws_client, snapshot, create_connection, auth_params, connection_name
-    ):
-        response = create_connection(
-            auth_params["AuthorizationType"],
-            auth_params["AuthParameters"],
-        )
-        snapshot.match("create-connection-auth", response)
-
-        describe_response = aws_client.events.describe_connection(Name=connection_name)
-        snapshot.match("describe-connection-auth", describe_response)
-
-    @markers.aws.validated
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    def test_list_connections(self, aws_client, snapshot, create_connection, connection_name):
-        create_connection(
-            "BASIC",
-            {
-                "BasicAuthParameters": {"Username": "user", "Password": "pass"},
-                "InvocationHttpParameters": {},
-            },
-        )
-
-        response = aws_client.events.list_connections(NamePrefix=connection_name)
-        snapshot.match("list-connections", response)
-
-    @markers.aws.validated
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    def test_delete_connection(self, aws_client, snapshot, create_connection, connection_name):
-        response = create_connection(
-            "API_KEY",
-            {
-                "ApiKeyAuthParameters": {"ApiKeyName": "ApiKey", "ApiKeyValue": "secret"},
-                "InvocationHttpParameters": {},
-            },
-        )
-        snapshot.match("create-connection-response", response)
-
-        secret_arn = aws_client.events.describe_connection(Name=connection_name)["SecretArn"]
-        # check if secret exists
-        aws_client.secretsmanager.describe_secret(SecretId=secret_arn)
-
-        delete_response = aws_client.events.delete_connection(Name=connection_name)
-        snapshot.match("delete-connection", delete_response)
-
-        # wait until connection is deleted
-        def is_connection_deleted():
-            try:
-                aws_client.events.describe_connection(Name=connection_name)
-                return False
-            except Exception:
-                return True
-
-        poll_condition(is_connection_deleted)
-
-        with pytest.raises(aws_client.events.exceptions.ResourceNotFoundException) as exc:
-            aws_client.events.describe_connection(Name=connection_name)
-        snapshot.match("describe-deleted-connection", exc.value.response)
-
-        def is_secret_deleted():
-            try:
-                aws_client.secretsmanager.describe_secret(SecretId=secret_arn)
-                return False
-            except Exception:
-                return True
-
-        poll_condition(is_secret_deleted)
-
-        with pytest.raises(aws_client.secretsmanager.exceptions.ResourceNotFoundException):
-            aws_client.secretsmanager.describe_secret(SecretId=secret_arn)
-
-    @markers.aws.validated
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    def test_create_connection_invalid_parameters(self, aws_client, snapshot, connection_name):
-        with pytest.raises(ClientError) as e:
-            aws_client.events.create_connection(
-                Name=connection_name,
-                AuthorizationType="INVALID_AUTH_TYPE",
-                AuthParameters={},
-            )
-        snapshot.match("create-connection-invalid-auth-error", e.value.response)
-
-    @markers.aws.validated
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    def test_update_connection(self, aws_client, snapshot, create_connection, connection_name):
-        create_response = create_connection(
-            "BASIC",
-            {
-                "BasicAuthParameters": {"Username": "user", "Password": "pass"},
-                "InvocationHttpParameters": {},
-            },
-        )
-        snapshot.match("create-connection", create_response)
-
-        describe_response = aws_client.events.describe_connection(Name=connection_name)
-        snapshot.match("describe-created-connection", describe_response)
-
-        # add secret id transformer
-        secret_id = describe_response["SecretArn"]
-        secret_uuid, _, secret_suffix = secret_id.rpartition("/")[2].rpartition("-")
-        snapshot.add_transformer(
-            snapshot.transform.regex(secret_uuid, "<secret-uuid>"), priority=-1
-        )
-        snapshot.add_transformer(
-            snapshot.transform.regex(secret_suffix, "<secret-id-suffix>"), priority=-1
-        )
-
-        get_secret_response = aws_client.secretsmanager.get_secret_value(SecretId=secret_id)
-        snapshot.match("connection-secret-before-update", get_secret_response)
-
-        update_response = aws_client.events.update_connection(
-            Name=connection_name,
-            AuthorizationType="BASIC",
-            AuthParameters={
-                "BasicAuthParameters": {"Username": "new_user", "Password": "new_pass"},
-                "InvocationHttpParameters": {},
-            },
-        )
-        snapshot.match("update-connection", update_response)
-
-        describe_response = aws_client.events.describe_connection(Name=connection_name)
-        snapshot.match("describe-updated-connection", describe_response)
-
-        get_secret_response = aws_client.secretsmanager.get_secret_value(SecretId=secret_id)
-        snapshot.match("connection-secret-after-update", get_secret_response)
-
-    @markers.aws.validated
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    def test_create_connection_name_validation(self, aws_client, snapshot, connection_name):
-        invalid_name = "Invalid Name With Spaces!"
-
-        with pytest.raises(ClientError) as e:
-            aws_client.events.create_connection(
-                Name=invalid_name,
-                AuthorizationType="API_KEY",
-                AuthParameters={
-                    "ApiKeyAuthParameters": {"ApiKeyName": "ApiKey", "ApiKeyValue": "secret"},
-                    "InvocationHttpParameters": {},
-                },
-            )
-        snapshot.match("create-connection-invalid-name-error", e.value.response)
-
-    @markers.aws.validated
-    @pytest.mark.parametrize(
-        "auth_params", API_DESTINATION_AUTH_PARAMS, ids=["basic", "api-key", "oauth"]
-    )
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    def test_connection_secrets(
-        self, aws_client, snapshot, create_connection, connection_name, auth_params
-    ):
-        response = create_connection(
-            auth_params["AuthorizationType"],
-            auth_params["AuthParameters"],
-        )
-        snapshot.match("create-connection-auth", response)
-
-        describe_response = aws_client.events.describe_connection(Name=connection_name)
-        snapshot.match("describe-connection-auth", describe_response)
-
-        secret_id = describe_response["SecretArn"]
-        secret_uuid, _, secret_suffix = secret_id.rpartition("/")[2].rpartition("-")
-        snapshot.add_transformer(
-            snapshot.transform.regex(secret_uuid, "<secret-uuid>"), priority=-1
-        )
-        snapshot.add_transformer(
-            snapshot.transform.regex(secret_suffix, "<secret-id-suffix>"), priority=-1
-        )
-        get_secret_response = aws_client.secretsmanager.get_secret_value(SecretId=secret_id)
-        snapshot.match("connection-secret", get_secret_response)
-
-
-API_DESTINATION_AUTHS = [
-    {
-        "type": "BASIC",
-        "key": "BasicAuthParameters",
-        "parameters": {"Username": "user", "Password": "pass"},
-    },
-    {
-        "type": "API_KEY",
-        "key": "ApiKeyAuthParameters",
-        "parameters": {"ApiKeyName": "ApiKey", "ApiKeyValue": "secret"},
-    },
-    {
-        "type": "OAUTH_CLIENT_CREDENTIALS",
-        "key": "OAuthParameters",
-        "parameters": {
-            "ClientParameters": {"ClientID": "id", "ClientSecret": "password"},
-            "AuthorizationEndpoint": "https://example.com/oauth",
-            "HttpMethod": "POST",
-            "OAuthHttpParameters": {
-                "BodyParameters": [{"Key": "oauthbody", "Value": "value1", "IsValueSecret": False}],
-                "HeaderParameters": [
-                    {"Key": "oauthheader", "Value": "value2", "IsValueSecret": False}
-                ],
-                "QueryStringParameters": [
-                    {"Key": "oauthquery", "Value": "value3", "IsValueSecret": False}
-                ],
-            },
-        },
-    },
-]
-
-
-class TestEventBridgeApiDestinations:
-    @pytest.fixture
-    def api_destination_snapshots(self, snapshot, destination_name):
-        """Common snapshot transformers for API destination tests."""
-        return TransformerUtility.eventbridge_api_destination(snapshot, destination_name)
-
-    @markers.aws.validated
-    @pytest.mark.parametrize("auth", API_DESTINATION_AUTHS)
-    @pytest.mark.skipif(
-        is_old_provider(),
-        reason="V1 provider does not support this feature",
-    )
-    def test_api_destinations(
-        self,
-        aws_client,
-        api_destination_snapshots,
-        create_connection,
-        create_api_destination,
-        connection_name,
-        destination_name,
-        auth,
-    ):
-        connection_response = create_connection(auth)
-        connection_arn = connection_response["ConnectionArn"]
-
-        response = create_api_destination(
-            ConnectionArn=connection_arn,
-            HttpMethod="POST",
-            InvocationEndpoint="https://example.com/api",
-            Description="Test API destination",
-        )
-        api_destination_snapshots.match("create-api-destination", response)
-
-        describe_response = aws_client.events.describe_api_destination(Name=destination_name)
-        api_destination_snapshots.match("describe-api-destination", describe_response)
-
-        list_response = aws_client.events.list_api_destinations(NamePrefix=destination_name)
-        api_destination_snapshots.match("list-api-destinations", list_response)
-
-        update_response = aws_client.events.update_api_destination(
-            Name=destination_name,
-            ConnectionArn=connection_arn,
-            HttpMethod="PUT",
-            InvocationEndpoint="https://example.com/api/v2",
-            Description="Updated API destination",
-        )
-        api_destination_snapshots.match("update-api-destination", update_response)
-
-        describe_updated_response = aws_client.events.describe_api_destination(
-            Name=destination_name
-        )
-        api_destination_snapshots.match(
-            "describe-updated-api-destination", describe_updated_response
-        )
-
-        delete_response = aws_client.events.delete_api_destination(Name=destination_name)
-        api_destination_snapshots.match("delete-api-destination", delete_response)
-
-        with pytest.raises(aws_client.events.exceptions.ResourceNotFoundException) as exc_info:
-            aws_client.events.describe_api_destination(Name=destination_name)
-        api_destination_snapshots.match(
-            "describe-api-destination-not-found-error", exc_info.value.response
-        )
-
-    @markers.aws.validated
-    @pytest.mark.skipif(is_old_provider(), reason="V1 provider does not support this feature")
-    def test_create_api_destination_invalid_parameters(
-        self, aws_client, api_destination_snapshots, connection_name, destination_name
-    ):
-        with pytest.raises(ClientError) as e:
-            aws_client.events.create_api_destination(
-                Name=destination_name,
-                ConnectionArn="invalid-connection-arn",
-                HttpMethod="INVALID_METHOD",
-                InvocationEndpoint="invalid-endpoint",
-            )
-        api_destination_snapshots.match(
-            "create-api-destination-invalid-parameters-error", e.value.response
-        )
-
-    @markers.aws.validated
-    @pytest.mark.skipif(is_old_provider(), reason="V1 provider does not support this feature")
-    def test_create_api_destination_name_validation(
-        self, aws_client, api_destination_snapshots, create_connection, connection_name
-    ):
-        invalid_name = "Invalid Name With Spaces!"
-
-        connection_response = create_connection(API_DESTINATION_AUTHS[0])
-        connection_arn = connection_response["ConnectionArn"]
-
-        with pytest.raises(ClientError) as e:
-            aws_client.events.create_api_destination(
-                Name=invalid_name,
-                ConnectionArn=connection_arn,
-                HttpMethod="POST",
-                InvocationEndpoint="https://example.com/api",
-            )
-        api_destination_snapshots.match(
-            "create-api-destination-invalid-name-error", e.value.response
         )
